Restore auth UI and consolidate hooks
